lib/sound: Fix oggopus callback type mismatch error
Move opus callbacks to oggopus.cpp and use the opus type defines.
